requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioresponses==0.7.8
aiosignal==1.3.2
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Flask==2.3.3
fonttools==4.57.0
frozenlist==1.5.0
gunicorn==21.2.0
idna==3.10
inference-sdk==0.9.17
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
propcache==0.3.1
py-cpuinfo==9.0.0
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.31.0
scipy==1.15.2
six==1.17.0
SpeechRecognition==3.14.2
standard-aifc==3.13.0
standard-chunk==3.13.0
supervision==0.25.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.19.0
